bowfa

+++ src/mahoji/lib/abstracted_commands/minionKill/newMinionKill.ts
@@ -29,6 +29,8 @@ import {
 import { killsRemainingOnTask } from './calcTaskMonstersRemaining';
 import { type PostBoostEffect, postBoostEffects } from './postBoostEffects';
 import { speedCalculations } from './timeAndSpeed';
+import { getSimilarItems } from '../../../../lib/data/similarItems';
+import getOSItem from '../../../../lib/util/getOSItem';
 
 const newMinionKillReturnSchema = z.object({
 	duration: z.number().int().positive(),
@@ -199,23 +201,28 @@ export function newMinionKillCommand(args: MinionKillOptions) {
 	}
 
 	if (monster.projectileUsage?.required) {
-		if (!gearBank.gear.range.ammo?.item) {
-			return `You need range ammo equipped to kill ${monster.name}.`;
-		}
 		const rangeCheck = checkRangeGearWeapon(gearBank.gear.range);
 		if (typeof rangeCheck === 'string') {
 			return `Your range gear isn't right: ${rangeCheck}`;
 		}
-		const projectilesNeeded = monster.projectileUsage.calculateQuantity({ quantity });
-		speedDurationResult.updateBank.itemCostBank.add(rangeCheck.ammo.item, projectilesNeeded);
-		if (projectilesNeeded > rangeCheck.ammo.quantity) {
-			return `You need ${projectilesNeeded.toLocaleString()}x ${itemNameFromID(
-				rangeCheck.ammo.item
-			)} to kill ${quantity}x ${
-				monster.name
-			}, and you have ${rangeCheck.ammo.quantity.toLocaleString()}x equipped.`;
+		const usingBowfa = getSimilarItems(getOSItem('Bow of faerdhinen (c)').id).includes(rangeCheck.weapon.id);
+		if (!gearBank.gear.range.ammo?.item && !usingBowfa) {
+			return `You need range ammo equipped to kill ${monster.name}.`;
+		}
+		
+		const projectilesNeeded = usingBowfa ? 0 : monster.projectileUsage.calculateQuantity({ quantity });
+		if (rangeCheck.ammo) {
+			speedDurationResult.updateBank.itemCostBank.add(rangeCheck.ammo.item, projectilesNeeded);
+			if (projectilesNeeded > rangeCheck.ammo.quantity) {
+				return `You need ${projectilesNeeded.toLocaleString()}x ${itemNameFromID(
+					rangeCheck.ammo.item
+				)} to kill ${quantity}x ${monster.name}, and you have ${rangeCheck.ammo.quantity.toLocaleString()}x equipped.`;
+			}
+		} else if (!usingBowfa) {
+			return `You need ammo equipped to kill ${monster.name}.`;
 		}
 	}
+	
 
 	for (const effect of [...postBoostEffects, ...ephemeralPostTripEffects]) {
 		const result = effect.run({
